Create new rename relation dropdown selection (#1284)
## Summary
Currently the relation dropdown, if you edit the string, gives you two
options: "Create new Relation X" or "Delete relation". We want to have
an option to edit the current relation.

My change adds a new option "Renames relation X to Y" which also
auto-updates the reverse node's string.

## How to test
Create a relation using double semicolons.
See that the reverse relation under __user_relation_types__ is there
Click on the relation from the original node, and type in a new string,
it should now show a new option for renaming.
Click on the rename option, and see that the relation is renamed
Check the reverse relation should be the new name + "of"

## Checklist

- [ ] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [x] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).