Non editable node visual v5 (#1280)
Added styling adjustments to the chip design Eric's PR
https://github.com/IdeaFlowCo/mew/pull/1263

Kept the icon, made the text black and added a gray background that
turns darker on hover.
The text cursor can no longer be placed on non editable nodes (not sure
if the way this is implemented is the best).
Hovering anywhere on the node highlights the chip.
Clicking anywhere on the node expands/collapses the node.

Disallowing placing the text cursor and the hand mouse pointer on hover
reduce the confusion for the user.

Added the chip styling to __user_relation_types__ and __card_statuses__
nodes as well.


![image](https://github.com/user-attachments/assets/e20d15ca-5de9-42a0-a9a9-c42f1d94d07f)

---------

Co-authored-by: Eric Cheng <ericcheng95@gmail.com>
Co-authored-by: Samuel Xu <samuelzxu@gmail.com>