Fix bug with relation selection closing early (#1287)
## Summary
There's an issue right now where when you open the relation selector,
close it, then reopen it, if you then highlight over the text, it'll
auto-close the selector.
This only happens the second time.

The cause of this is that onCloseAutoFocus is automatically triggered
whenever an "external" event is handled. And highlighting the input is
considered external.

To fix this we need to do two things, firstly override onInteractOutside
to override e.preventDefault. This makes it so that clicking outside no
longer auto-triggers the popover close. We have to also add a
inputRef.current.focus because otherwise it'll remove the highlight as
well.

We then add a new effect, which will check if the mouse click is
external from the input. If it is, then close the popover.

## How to test
Open a relation selection (if you don't have one, create a new one and
then refresh the page)
Highlight the input, it should not close
Click outside of the box, the selection should close now

Do the same thing again, and verify that it still works.

## Checklist

- [ ] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [x] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).