Fix double bracket mention bug (#1292)
## Summary
There's a bug where if you use multiple double brackets for the node,
each time you add it, it gets added as a child relation below.

https://www.loom.com/share/e7db7941e5cf4ba8b734236f4db0ef59

This was happening because the double bracket relations are slightly
different than the other relations (graphNode and treeNode to and from
and flipped):
 await graphStore.addRelation({
    fromId: graphNodeId,
    toId: treeNode.object.id,
    relationTypeId: ...,
  });
  
Because of this, hasMentionOrParentRelation isn't detecting that it
already exists.
I fixed this by creating a new path for double brackets, and it uses a
new algorithm hasDoubleBracketRelation instead.

## How to test
Enable Roam Research setting
Create a double bracket mention.
Observe that the relation is below it
Create another double bracket mention.
Observe that there should only be one mention still.
Try the other mentions, this shouldn't break any of those existing ones.

## Checklist

- [ ] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [x] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).