Fix bug where double brackets can trigger TODO checkbox (#1293)
## Summary
If you use double square brackets for a mention, and try to close the
bracket with a square bracket, the parser considers it a TODO input
(since it starts with [ and ends with ]).

My fix checks to make sure that the string doesn't start with a double
bracket first.

## How to test
Create a new node
Open double brackets "[[" + some node name
Close the double brackets with a single "]"
It shouldn't create a TODO checkbox.

Enable the Roam Research setting and do the same thing
It should create mention, but not a TODO checkbox.

## Checklist

- [ ] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [x] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).