Update isTextSame to account for bold/italics changes (#1303)
## Summary

https://www.loom.com/share/9648913f3d6747c2837f2ad238bf7556?sid=46003501-79cc-400a-a5d2-365919050513
Bolding/Italicizing is currently classified as the same text, so it's
appended in the undoStack's update. This makes it so that if you do
ctrl-z, it undos both the bold/italicizing and the previous change.

I updated the IsTextSame function to return false if the last change
changed the bold/italicization.

## How to test

Provide steps on how to test the fix/feature (which you know, because
you tested it 👀)

## Checklist

- [ ] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [x] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).