Clicking in the empty area in the bottom places the cursor on a new node (#1311)
## Summary

This PR updates the click behavior for empty space at the bottom of
pages to focus on the last node _only if it is empty_ and to focus on a
new node otherwise.

This fixes cursor position issues in cases like
- having uneditable chip nodes at the end of the page (like in the
landing page of a new dev instance)
- having transcluded nodes (ones with boxes) at the end of the page
(where the cursor is barely noticeable at the end of the box and typing
starts appending the transcluded node)

Completes
[ENT-4878](https://linear.app/ideaflow/issue/ENT-4878/click-on-empty-space-to-place-cursor-on-new-node)

## Changes

**OutlineContent.tsx**
- Enhanced the `EmptySpaceClickArea` onClick handler with empty node
detection
- Replaced focus on last node logic to check if it is empty (no text
content and no children)
   - Focus on the last visible node only if it is empty
   - Create a new node when there's no empty node at the bottom

## How to test

1. Navigate to any page in the application
2. Click anywhere in the empty space at the bottom of the page (below
the last node)
3. Verify that a new node is created at the bottom with the cursor
positioned in it
4. Test this behavior on pages with different types of content in the
last node:
   - with regular text
   - with transcluded nodes
   - with nested/expanded nodes
- with nothing (empty node) - in this case, clicking in the empty space
should focus on this node
5. Click multiple times in the empty space and verify that it does not
create a new node every time, but focuses on the last empty node

## Checklist

- [x] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [x] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).

---------

Co-authored-by: Samuel Xu <samuelzxu@gmail.com>