Update test path so that commands work (#1308)
## Summary
The current path for the tests doesn't seem to work for my computer
(Windows). This change should make it so that both Mac and Windows
works.

## How to test
Run:
npx jest --listTests
It should return a list of test files. If it returns nothing then it's
not working.

## Checklist

- [ ] I followed the [style
guide](https://www.notion.so/Style-guide-08041425021b49b28a0734f4e8105c33).
- [x] I tested my contribution.
- [ ] I added new environment variables to
[Vercel](https://vercel.com/ideaflowco/mew/settings/environment-variables).