fix: stop initialising with unlogged user, wait until we have an user (#1326)
Previously, we initialized stores with an “unlogged” user because we had
to wait for useSetupUser. We introduced a `UserProvider` and lifted the
user setup logic. This removes the dual responsibility from
StoreProvider, as in, it no longer needs to initialize the user and is
now solely responsible for setting up stores. It also removes reliance
on the temporary "unloged" user state and guarantees that a user exists
before StoreProvider is rendered.